requirements.txt
torch
torchvision
timm
pyyaml
tqdm
numpy
